+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/attachments/factory/AttachmentsPickerImagesTabFactory.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.compose.ui.messages.attachments.facotry
+package io.getstream.chat.android.compose.ui.messages.attachments.factory
 
 import android.Manifest
 import androidx.compose.foundation.layout.padding
