+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/messages/attachments/factory/AttachmentsPickerTabFactories.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.getstream.chat.android.compose.ui.messages.attachments.facotry
+package io.getstream.chat.android.compose.ui.messages.attachments.factory
 
 public object AttachmentsPickerTabFactories {
